requirements.txt
pytelegrambotapi
python-telegram-bot
pathlib
typing
openai
loguru
pydantic
pymongo

brunette
isort
autopep8
autoflake
pylint
flake8
loguru